+++ circuit_maintenance_parser/__init__.py
@@ -1,9 +1,10 @@
 """Notifications parser init."""
 
 from typing import Type, Optional, Iterable
+from pydantic import BaseModel
 
 from .errors import NonexistentParserError, ParsingError
-
+from .output import Maintenance
 from .providers import (
     GenericProvider,
     Cogent,
@@ -32,7 +33,7 @@ SUPPORTED_PROVIDER_NAMES = [provider.get_provider_type() for provider in SUPPORT
 SUPPORTED_ORGANIZER_EMAILS = [provider.get_default_organizer() for provider in SUPPORTED_PROVIDERS]
 
 
-def init_parser(raw, provider_type=None) -> Optional[GenericProvider]:
+def init_provider(raw, provider_type=None) -> Optional[GenericProvider]:
     """Returns an instance of the corresponding Notification Parser."""
     try:
         if not provider_type:
@@ -88,8 +89,56 @@ def get_provider_data_types(provider_name: str) -> Iterable[str]:
     return provider.get_data_types()
 
 
+# TODO: In Release 1.2.0, we introduced new Provider abstraction to accomodate multiple potential parser per provider.
+# In order to mantain backwards compatibility until a new major version release (2.x.x) we still mantain the following
+# aliases that will be deprecated with the next major version change.
+
+
+class MaintenanceNotification(BaseModel):
+    """Class to keep backwards compatibility after changes in 1.2.0."""
+
+    raw: bytes
+    provider_type: str
+    sender: str = ""
+    subject: str = ""
+    source: str = ""
+
+    provider: Optional[GenericProvider] = None
+
+    def process(self) -> Iterable[Maintenance]:
+        """Method that returns a list of Maintenance objects."""
+        if self.provider:
+            return self.provider.process()
+        return []
+
+
+def init_parser(**kwargs) -> Optional[MaintenanceNotification]:
+    """Returns an instance of the MaintenanceNotification for backwards compatibility."""
+    try:
+        provider_type = kwargs.get("provider_type")
+        if not provider_type:
+            provider_type = GenericProvider.get_provider_type()
+        provider_parser_class = get_provider_class(provider_type)
+        return_value = MaintenanceNotification(**kwargs)
+        return_value.provider = provider_parser_class(raw=kwargs.get("raw"))
+        return return_value
+
+    except NonexistentParserError:
+        return None
+
+
+def get_provider_data_type(provider_name: str) -> Optional[str]:
+    """Function to return only the first available data_type."""
+    for data_type in get_provider_data_types(provider_name):
+        return data_type
+    return None
+
+
+# End of backwards compatibility section for Release 1.2.0
+
+
 __all__ = [
-    "init_parser",
+    "init_provider",
     "get_provider_class",
     "get_provider_class_from_sender",
     "get_provider_data_types",